% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_func.R
\name{food_func}
\alias{food_func}
\title{Food function}
\usage{
food_func(food, object)
}
\arguments{
\item{food}{insert a type of food as a character}

\item{object}{insert an object as a character}
}
\value{

}
\description{
Food function
}
