% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_func.R
\name{hello_func}
\alias{hello_func}
\title{Random Func}
\usage{
hello_func(name, year, candy)
}
\arguments{
\item{name}{a character containing the input name}

\item{year}{a year (number)}

\item{candy}{a character}
}
\value{

}
\description{
Random Func
}
