% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaning_of_life_func.R
\name{your_meaning_of_life_function}
\alias{your_meaning_of_life_function}
\title{Meaning of life function}
\usage{
your_meaning_of_life_function(age, height)
}
\arguments{
\item{age}{insert numeric}

\item{height}{insert numeric in cm}
}
\value{

}
\description{
Meaning of life function
}
